review: marsh quest markers + tighten review-comment guards
Marsh "Preparation" quest UX — visual cues across the full P90 → Marsh →
Terminal → Rings sequence:

* Chain 1055 (P90 pickup) re-sets Marsh's mission-available marker
  (mask 8388608) so the player sees a clear "talk to me next" cue
  for step 3563. Cleared in chains 1058/1059 once the briefing
  completes.
* Chains 1058/1059 (briefing dialog accept) set the same
  INT_MissionWorldObject quest highlight (mask 1073741824) on the
  Terminal that Marsh sends the player to next, mirroring the
  treatment we already give the SMG1A locker and the Ambernol vial.
  Cleared in chain 1060 the moment the player engages the Livewire
  minigame at the terminal.
* Chain 1061 (Livewire victory) sets the same quest highlight on
  Preparation_RingSwitch in addition to the existing RingNetwork
  bit (mask 32) — visual cue for "go here next" carried into
  mission 680. Cleared in chain 1072 on teleport_in.
* Chain 1074 (player_loaded restore for the second ring switch)
  now also re-applies the quest highlight on login alongside the
  RingNetwork bit it already restored.

New restore chains for restart-resilience (mirrors existing 1062
SMG1A / 1063 ColMarsh / 1074 ring-switch patterns):

* 1064 — player_loaded + step 3563 active → restore Marsh marker.
* 1065 — player_loaded + step 3564 active → restore Terminal
  quest highlight.

Review-comment fixes (PR #214 second round):

* CR `cell/service/tests.rs:258` — file crossed the 700-line hard
  cap from CLAUDE.md. Split into a `tests/` directory:
  `tests/mod.rs` (helper + module decls), `tests/reload.rs`,
  `tests/bandolier.rs`, `tests/npc_ai.rs`. All 12 tests pass under
  the new layout; helper lives as `pub(super) fn make_test_space_mgr`
  so submodules import via `super::`. Per-submodule line counts
  well under the 500-line soft cap.
* CR `player/world.rs:594` — strengthen
  `handle_reload_sends_ability_begin_sequence` from "first 4 bytes
  match" to full 26-byte ON_SEQUENCE wire-frame validation
  (sequence_id, source_id, target_id, primary, impact_time,
  nvp_count, view_type, instance_id) and pin `== 1` count instead
  of bool flag.
* CR `service/tests.rs:247` — same tightening on
  `reload_completion_tick_emits_ability_end_sequence_when_event_set_present`:
  exact 26-byte length, all fields validated, exact-one count.
* CR `chain_replay_tests.rs:194` — tighten the ambernol guard from
  `> 0` to `== 1` exact match so a future seed change that
  duplicates the remove action (causing a stack-of-2 vials to
  vanish in one use) fails the guard.
* Copilot `inventory/services.md`, `inventory/content-engine.md` —
  update per-crate totals (550→563, 63→64) and `Last updated` to
  match the README header. Resolves the per-crate vs README drift
  flagged in the review.

CI verification: fmt clean, clippy clean across the workspace,
563/563 services live-DB pass, 64/64 content-engine pass.
